chore: more assertions moved
